gitlab: call it fixtures also
