+++ server/forge/gitlab/fixtures/hooks.go
@@ -12,11 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package testdata
+package fixtures
 
 import (
 	"net/http"
 	"net/url"
+
+	_ "embed"
 )
 
 var (
