+++ server/forge/gitlab/fixtures/oauth.go
@@ -12,27 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package testdata
+package fixtures
 
-var currentUserPayload = []byte(`
-{
-  "id": 1,
-  "username": "john_smith",
-  "email": "john@example.com",
-  "name": "John Smith",
-  "private_token": "dd34asd13as",
-  "state": "active",
-  "created_at": "2012-05-23T08:00:58Z",
-  "bio": null,
-  "skype": "",
-  "linkedin": "",
-  "twitter": "",
-  "website_url": "",
-  "theme_id": 1,
-  "color_scheme_id": 2,
-  "is_admin": false,
-  "can_create_group": true,
-  "can_create_project": true,
-  "projects_limit": 100
-}
-`)
+var accessTokenPayload = []byte(`access_token=sekret&scope=api&token_type=bearer`)
