+++ server/forge/gitlab/fixtures/testdata.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package testdata
+package fixtures
 
 import (
 	"net/http"
